// src/graphql/generated/types.tsx
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DateTime: any;
  Upload: any;
};
// ...
export type Chats = {
  __typename?: "Chats";
  chatId: Scalars["String"];
  message: Message;
  roomId: PartialRoom;
  sender: PartialUser;
  toUser?: Maybe<PartialUser>;
};
// ...
export type Message = {
  __typename?: "Message";
  created_at: Scalars["Float"];
  duration?: Maybe<Scalars["Float"]>;
  fileURL: Scalars["String"];
  fontStyle: Scalars["String"];
  index: Scalars["Float"];
  isForwarded: Scalars["Boolean"];
  isSent: Scalars["Boolean"];
  message: Scalars["String"];
  thumbnail: Scalars["String"];
  type: Scalars["String"];
  updated_at: Scalars["Float"];
};
// ...
export type PartialRoom = {
  __typename?: "partialRoom";
  _id: Scalars["String"];
  name: Scalars["String"];
  type: Scalars["String"];
};

export type PartialUser = {
  __typename?: "partialUser";
  _id: Scalars["String"];
  firstName: Scalars["String"];
  lastName: Scalars["String"];
  profile_img: Scalars["String"];
};